improved Math lib

+++ Mesh/IntersectionData.cpp
@@ -8,5 +8,5 @@
 IntersectionData::IntersectionData():
         t( __FLT_MAX__ ), primitive( nullptr ) {}
 
-IntersectionData::IntersectionData( float t, Primitive* prim ):
+IntersectionData::IntersectionData( double t, Primitive* prim ):
         t( t ), primitive( prim ) {}
